fix: Same fix as last commit but for integrated servers as well.

+++ src/main/java/com/gildedgames/orbis/client/gui/GuiViewProjects.java
@@ -3,6 +3,7 @@ package com.gildedgames.orbis.client.gui;
 import com.gildedgames.aether.api.io.NBTFunnel;
 import com.gildedgames.aether.api.orbis.IWorldObject;
 import com.gildedgames.aether.api.orbis.exceptions.OrbisMissingProjectException;
+import com.gildedgames.aether.api.orbis.management.IData;
 import com.gildedgames.aether.api.orbis.management.IProject;
 import com.gildedgames.aether.api.orbis.management.IProjectIdentifier;
 import com.gildedgames.aether.common.AetherCore;
@@ -159,18 +160,32 @@ public class GuiViewProjects extends GuiFrame implements IDirectoryNavigatorList
 				{
 					final IWorldObject worldObject = this.blueprint;
 
-					if (this.project != null && worldObject.getData() != null)
+					if (this.project != null && worldObject.getData() != null && !file.exists())
 					{
-						worldObject.getData().preSaveToDisk(worldObject);
+						IData data = worldObject.getData();
 
-						this.project.getCache().setData(worldObject.getData(), location);
+						/**
+						 * Check if the data has already been stored.
+						 * If so, we should create a new identifier for it as
+						 * a clone. Many issues are caused if two files use
+						 * the same identifier.
+						 */
+						if (data.getMetadata().getIdentifier() != null && this.project.getCache().hasData(data.getMetadata().getIdentifier().getDataId()))
+						{
+							data = data.clone();
+							data.getMetadata().setIdentifier(this.project.getCache().createNextIdentifier());
+						}
+
+						data.preSaveToDisk(worldObject);
+
+						this.project.getCache().setData(data, location);
 
 						try (FileOutputStream out = new FileOutputStream(file))
 						{
 							final NBTTagCompound tag = new NBTTagCompound();
 							final NBTFunnel funnel = AetherCore.io().createFunnel(tag);
 
-							funnel.set("data", worldObject.getData());
+							funnel.set("data", data);
 
 							CompressedStreamTools.writeCompressed(tag, out);
